Add (new) core candidate filter to dbscan
The core candidate filter does not function the way I had implemented it for
IV19. The difference is that core points are filtered after finding nearest
neighbors and not before. As a result, also points which are not core
candidates contribute count as neighbors and thus may promote other
points to core points which would not have been core points before.